# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Task Manager API",
        "description": "API для управления задачами и пунктами задач",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "API Server"
        }
    ],
    "paths": {
        "/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Авторизация пользователя",
                "operationId": "d55f416d61f094eb6038b566bc6e4596",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "302": {
                        "description": "Перенаправление на список задач"
                    },
                    "422": {
                        "description": "Ошибка валидации"
                    }
                }
            }
        },
        "/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Регистрация нового пользователя",
                "operationId": "c954ad7948ee6a4792ee061c79906fc7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Иван"
                                    },
                                    "surname": {
                                        "type": "string",
                                        "example": "Иванов"
                                    },
                                    "patronymic": {
                                        "type": "string",
                                        "example": "Иванович"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "302": {
                        "description": "Перенаправление на список задач"
                    },
                    "422": {
                        "description": "Ошибка валидации"
                    }
                }
            }
        },
        "/tasks": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Получить список задач текущего пользователя",
                "operationId": "6a93d54730cde509d735e538ab6800f6",
                "responses": {
                    "200": {
                        "description": "Список задач с пунктами",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "task": {
                                                "type": "string",
                                                "example": "Купить продукты"
                                            },
                                            "user_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "task_points": {
                                                "type": "array",
                                                "items": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "task_point": {
                                                            "type": "string",
                                                            "example": "Купить хлеб"
                                                        },
                                                        "task_id": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "status": {
                                                            "type": "boolean",
                                                            "example": false
                                                        },
                                                        "created_at": {
                                                            "type": "string",
                                                            "format": "date-time"
                                                        },
                                                        "updated_at": {
                                                            "type": "string",
                                                            "format": "date-time"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Создать новую задачу",
                "operationId": "f22166d8ef8f7035d351d1180d1bdc82",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "task"
                                ],
                                "properties": {
                                    "task": {
                                        "type": "string",
                                        "example": "Купить продукты"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "302": {
                        "description": "Перенаправление на список задач"
                    },
                    "422": {
                        "description": "Ошибка валидации"
                    }
                }
            }
        },
        "/tasks/{id}": {
            "put": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Редактировать задачу",
                "operationId": "89561ac92b8b503ccc857dffe9bd28b5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID задачи",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "task_id",
                                    "task"
                                ],
                                "properties": {
                                    "task_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "task": {
                                        "type": "string",
                                        "example": "Обновлённая задача"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "302": {
                        "description": "Перенаправление на список задач"
                    },
                    "404": {
                        "description": "Задача не найдена"
                    },
                    "422": {
                        "description": "Ошибка валидации"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Удалить задачу",
                "operationId": "df3619c55f924fe386902218ba5983a7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID задачи",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Перенаправление на список задач"
                    },
                    "404": {
                        "description": "Задача не найдена"
                    }
                }
            }
        },
        "/task-points": {
            "post": {
                "tags": [
                    "TaskPoints"
                ],
                "summary": "Создать пункт задачи",
                "operationId": "64fd987e8b96d90a6373cdf620676b60",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "task_point",
                                    "task_id"
                                ],
                                "properties": {
                                    "task_point": {
                                        "type": "string",
                                        "example": "Купить хлеб"
                                    },
                                    "task_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "302": {
                        "description": "Перенаправление на список задач"
                    },
                    "422": {
                        "description": "Ошибка валидации"
                    }
                }
            }
        },
        "/task-points/{id}": {
            "put": {
                "tags": [
                    "TaskPoints"
                ],
                "summary": "Редактировать пункт задачи",
                "operationId": "71a96c80b41c4d45075bfc770aee2604",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID пункта задачи",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "task_point",
                                    "task_id"
                                ],
                                "properties": {
                                    "task_point": {
                                        "type": "string",
                                        "example": "Купить батон"
                                    },
                                    "task_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "302": {
                        "description": "Перенаправление на список задач"
                    },
                    "404": {
                        "description": "Пункт не найден"
                    },
                    "422": {
                        "description": "Ошибка валидации"
                    }
                }
            },
            "delete": {
                "tags": [
                    "TaskPoints"
                ],
                "summary": "Удалить пункт задачи",
                "operationId": "da8a88616d848bc64590db5e681830f6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID пункта задачи",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Перенаправление на список задач"
                    },
                    "404": {
                        "description": "Пункт не найден"
                    }
                }
            }
        },
        "/task-points/{id}/status": {
            "patch": {
                "tags": [
                    "TaskPoints"
                ],
                "summary": "Изменить статус пункта задачи",
                "operationId": "489034ffed2f80d49191c2336c24f9b0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID пункта задачи",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status"
                                ],
                                "properties": {
                                    "status": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Статус обновлён",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "Success"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "task_point": {
                                                    "type": "string",
                                                    "example": "Купить хлеб"
                                                },
                                                "task_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "status": {
                                                    "type": "boolean",
                                                    "example": true
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Пункт не найден"
                    },
                    "422": {
                        "description": "Ошибка валидации"
                    }
                }
            }
        },
        "/user": {
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "Обновить данные текущего пользователя",
                "operationId": "49d38ea7fe04e833b5be3c31258fa551",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "surname",
                                    "name",
                                    "email"
                                ],
                                "properties": {
                                    "surname": {
                                        "type": "string",
                                        "example": "Иванов"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Иван"
                                    },
                                    "patronymic": {
                                        "type": "string",
                                        "example": "Иванович"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "updated@example.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "302": {
                        "description": "Перенаправление на страницу аккаунта"
                    },
                    "404": {
                        "description": "Пользователь не найден"
                    },
                    "422": {
                        "description": "Ошибка валидации"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Authentication",
            "description": "Вход и регистрация"
        },
        {
            "name": "Tasks",
            "description": "Управление задачами"
        },
        {
            "name": "TaskPoints",
            "description": "Управление пунктами задач"
        },
        {
            "name": "User",
            "description": "Управление профилем пользователя"
        }
    ]
}